// src/BindingChaos.Web/src/api/models/UserGroupCharterResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserGroupShunningRulesResponse } from './UserGroupShunningRulesResponse';
import {
    UserGroupShunningRulesResponseFromJSON,
    UserGroupShunningRulesResponseFromJSONTyped,
    UserGroupShunningRulesResponseToJSON,
    UserGroupShunningRulesResponseToJSONTyped,
} from './UserGroupShunningRulesResponse';
import type { UserGroupMembershipRulesResponse } from './UserGroupMembershipRulesResponse';
import {
    UserGroupMembershipRulesResponseFromJSON,
    UserGroupMembershipRulesResponseFromJSONTyped,
    UserGroupMembershipRulesResponseToJSON,
    UserGroupMembershipRulesResponseToJSONTyped,
} from './UserGroupMembershipRulesResponse';
import type { UserGroupContestationRulesResponse } from './UserGroupContestationRulesResponse';
import {
    UserGroupContestationRulesResponseFromJSON,
    UserGroupContestationRulesResponseFromJSONTyped,
    UserGroupContestationRulesResponseToJSON,
    UserGroupContestationRulesResponseToJSONTyped,
} from './UserGroupContestationRulesResponse';

/**
 * Represents the charter of a user group.
 * @export
 * @interface UserGroupCharterResponse
 */
export interface UserGroupCharterResponse {
    /**
     * 
     * @type {UserGroupMembershipRulesResponse}
     * @memberof UserGroupCharterResponse
     */
    membershipRules?: UserGroupMembershipRulesResponse;
    /**
     * 
     * @type {UserGroupContestationRulesResponse}
     * @memberof UserGroupCharterResponse
     */
    contestationRules?: UserGroupContestationRulesResponse;
    /**
     * 
     * @type {UserGroupShunningRulesResponse}
     * @memberof UserGroupCharterResponse
     */
    shunningRules?: UserGroupShunningRulesResponse;
}

/**
 * Check if a given object implements the UserGroupCharterResponse interface.
 */
export function instanceOfUserGroupCharterResponse(value: object): value is UserGroupCharterResponse {
    return true;
}

export function UserGroupCharterResponseFromJSON(json: any): UserGroupCharterResponse {
    return UserGroupCharterResponseFromJSONTyped(json, false);
}

export function UserGroupCharterResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserGroupCharterResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'membershipRules': json['membershipRules'] == null ? undefined : UserGroupMembershipRulesResponseFromJSON(json['membershipRules']),
        'contestationRules': json['contestationRules'] == null ? undefined : UserGroupContestationRulesResponseFromJSON(json['contestationRules']),
        'shunningRules': json['shunningRules'] == null ? undefined : UserGroupShunningRulesResponseFromJSON(json['shunningRules']),
    };
}

export function UserGroupCharterResponseToJSON(json: any): UserGroupCharterResponse {
    return UserGroupCharterResponseToJSONTyped(json, false);
}

export function UserGroupCharterResponseToJSONTyped(value?: UserGroupCharterResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'membershipRules': UserGroupMembershipRulesResponseToJSON(value['membershipRules']),
        'contestationRules': UserGroupContestationRulesResponseToJSON(value['contestationRules']),
        'shunningRules': UserGroupShunningRulesResponseToJSON(value['shunningRules']),
    };
}