// src/BindingChaos.Web/src/api/models/UserGroupMembershipRulesResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserGroupApprovalSettingsResponse } from './UserGroupApprovalSettingsResponse';
import {
    UserGroupApprovalSettingsResponseFromJSON,
    UserGroupApprovalSettingsResponseFromJSONTyped,
    UserGroupApprovalSettingsResponseToJSON,
    UserGroupApprovalSettingsResponseToJSONTyped,
} from './UserGroupApprovalSettingsResponse';

/**
 * Represents the membership rules of a user group charter.
 * @export
 * @interface UserGroupMembershipRulesResponse
 */
export interface UserGroupMembershipRulesResponse {
    /**
     * The policy controlling how participants may join.
     * @type {string}
     * @memberof UserGroupMembershipRulesResponse
     */
    joinPolicy?: string;
    /**
     * The maximum number of members allowed, if set.
     * @type {number}
     * @memberof UserGroupMembershipRulesResponse
     */
    maxMembers?: number | null;
    /**
     * The entry requirements for joining, if any.
     * @type {string}
     * @memberof UserGroupMembershipRulesResponse
     */
    entryRequirements?: string | null;
    /**
     * Whether the member list is publicly visible.
     * @type {boolean}
     * @memberof UserGroupMembershipRulesResponse
     */
    memberListPublic?: boolean;
    /**
     * 
     * @type {UserGroupApprovalSettingsResponse}
     * @memberof UserGroupMembershipRulesResponse
     */
    approvalSettings?: UserGroupApprovalSettingsResponse | null;
}

/**
 * Check if a given object implements the UserGroupMembershipRulesResponse interface.
 */
export function instanceOfUserGroupMembershipRulesResponse(value: object): value is UserGroupMembershipRulesResponse {
    return true;
}

export function UserGroupMembershipRulesResponseFromJSON(json: any): UserGroupMembershipRulesResponse {
    return UserGroupMembershipRulesResponseFromJSONTyped(json, false);
}

export function UserGroupMembershipRulesResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserGroupMembershipRulesResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'joinPolicy': json['joinPolicy'] == null ? undefined : json['joinPolicy'],
        'maxMembers': json['maxMembers'] == null ? undefined : json['maxMembers'],
        'entryRequirements': json['entryRequirements'] == null ? undefined : json['entryRequirements'],
        'memberListPublic': json['memberListPublic'] == null ? undefined : json['memberListPublic'],
        'approvalSettings': json['approvalSettings'] == null ? undefined : UserGroupApprovalSettingsResponseFromJSON(json['approvalSettings']),
    };
}

export function UserGroupMembershipRulesResponseToJSON(json: any): UserGroupMembershipRulesResponse {
    return UserGroupMembershipRulesResponseToJSONTyped(json, false);
}

export function UserGroupMembershipRulesResponseToJSONTyped(value?: UserGroupMembershipRulesResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'joinPolicy': value['joinPolicy'],
        'maxMembers': value['maxMembers'],
        'entryRequirements': value['entryRequirements'],
        'memberListPublic': value['memberListPublic'],
        'approvalSettings': UserGroupApprovalSettingsResponseToJSON(value['approvalSettings']),
    };
}