// src/BindingChaos.Web/src/api/models/UserGroupContestationRulesResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Represents the contestation rules of a user group charter.
 * @export
 * @interface UserGroupContestationRulesResponse
 */
export interface UserGroupContestationRulesResponse {
    /**
     * The duration allowed for resolving a contestation.
     * @type {string}
     * @memberof UserGroupContestationRulesResponse
     */
    resolutionWindow?: string;
    /**
     * The fraction of rejections required to block a contested action.
     * @type {number}
     * @memberof UserGroupContestationRulesResponse
     */
    rejectionThreshold?: number;
}

/**
 * Check if a given object implements the UserGroupContestationRulesResponse interface.
 */
export function instanceOfUserGroupContestationRulesResponse(value: object): value is UserGroupContestationRulesResponse {
    return true;
}

export function UserGroupContestationRulesResponseFromJSON(json: any): UserGroupContestationRulesResponse {
    return UserGroupContestationRulesResponseFromJSONTyped(json, false);
}

export function UserGroupContestationRulesResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserGroupContestationRulesResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'resolutionWindow': json['resolutionWindow'] == null ? undefined : json['resolutionWindow'],
        'rejectionThreshold': json['rejectionThreshold'] == null ? undefined : json['rejectionThreshold'],
    };
}

export function UserGroupContestationRulesResponseToJSON(json: any): UserGroupContestationRulesResponse {
    return UserGroupContestationRulesResponseToJSONTyped(json, false);
}

export function UserGroupContestationRulesResponseToJSONTyped(value?: UserGroupContestationRulesResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'resolutionWindow': value['resolutionWindow'],
        'rejectionThreshold': value['rejectionThreshold'],
    };
}