// src/BindingChaos.Web/src/api/models/UserGroupShunningRulesResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Represents the shunning rules of a user group charter.
 * @export
 * @interface UserGroupShunningRulesResponse
 */
export interface UserGroupShunningRulesResponse {
    /**
     * The fraction of member votes required to shun a participant.
     * @type {number}
     * @memberof UserGroupShunningRulesResponse
     */
    approvalThreshold?: number;
}

/**
 * Check if a given object implements the UserGroupShunningRulesResponse interface.
 */
export function instanceOfUserGroupShunningRulesResponse(value: object): value is UserGroupShunningRulesResponse {
    return true;
}

export function UserGroupShunningRulesResponseFromJSON(json: any): UserGroupShunningRulesResponse {
    return UserGroupShunningRulesResponseFromJSONTyped(json, false);
}

export function UserGroupShunningRulesResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserGroupShunningRulesResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'approvalThreshold': json['approvalThreshold'] == null ? undefined : json['approvalThreshold'],
    };
}

export function UserGroupShunningRulesResponseToJSON(json: any): UserGroupShunningRulesResponse {
    return UserGroupShunningRulesResponseToJSONTyped(json, false);
}

export function UserGroupShunningRulesResponseToJSONTyped(value?: UserGroupShunningRulesResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'approvalThreshold': value['approvalThreshold'],
    };
}