// src/BindingChaos.Web/src/api/models/UserGroupMemberResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Represents a single member of a user group.
 * @export
 * @interface UserGroupMemberResponse
 */
export interface UserGroupMemberResponse {
    /**
     * The pseudonym of the member.
     * @type {string}
     * @memberof UserGroupMemberResponse
     */
    pseudonym?: string;
}

/**
 * Check if a given object implements the UserGroupMemberResponse interface.
 */
export function instanceOfUserGroupMemberResponse(value: object): value is UserGroupMemberResponse {
    return true;
}

export function UserGroupMemberResponseFromJSON(json: any): UserGroupMemberResponse {
    return UserGroupMemberResponseFromJSONTyped(json, false);
}

export function UserGroupMemberResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserGroupMemberResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'pseudonym': json['pseudonym'] == null ? undefined : json['pseudonym'],
    };
}

export function UserGroupMemberResponseToJSON(json: any): UserGroupMemberResponse {
    return UserGroupMemberResponseToJSONTyped(json, false);
}

export function UserGroupMemberResponseToJSONTyped(value?: UserGroupMemberResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'pseudonym': value['pseudonym'],
    };
}